♻️ refactor: components폴더 안에 pages/quiz폴더 생성
