+++ src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './Text';
+export * from './pages';
