+++ src/components/pages/quiz/Comment/LikeButton.tsx
@@ -5,7 +5,7 @@ import clsx from 'clsx';
 import Image from 'next/image';
 import { HTMLAttributes, useState } from 'react';
 
-import { Text } from '../Text';
+import { Text } from '../../../Text';
 
 interface LikeButtonProps
   extends Omit<HTMLAttributes<HTMLSpanElement>, 'className'> {
