+++ src/components/pages/quiz/index.ts
@@ -1,0 +1,1 @@
+export * from './Comment';
